Move predicates into a package
