+++ src/main/java/seedu/address/logic/commands/ViewCommand.java
@@ -9,7 +9,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.SamePersonPredicate;
+import seedu.address.model.person.predicates.SamePersonPredicate;
 
 /**
  * Views the details of an existing person in the address book.
