+++ src/main/java/seedu/address/model/ModelManager.java
@@ -12,7 +12,7 @@ import javafx.collections.transformation.FilteredList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.SamePersonPredicate;
+import seedu.address.model.person.predicates.SamePersonPredicate;
 
 /**
  * Represents the in-memory model of the address book data.
