+++ src/main/java/seedu/address/model/person/predicates/SamePersonPredicate.java
@@ -1,4 +1,6 @@
-package seedu.address.model.person;
+package seedu.address.model.person.predicates;
+
+import seedu.address.model.person.Person;
 
 import java.util.function.Predicate;
