+++ src/test/java/seedu/address/logic/commands/ViewCommandTest.java
@@ -17,7 +17,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.SamePersonPredicate;
+import seedu.address.model.person.predicates.SamePersonPredicate;
 
 public class ViewCommandTest {
     private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
