+++ src/test/java/seedu/address/model/person/SamePersonPredicateTest.java
@@ -21,6 +21,7 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.person.predicates.SamePersonPredicate;
 import seedu.address.testutil.PersonBuilder;
 
 public class SamePersonPredicateTest {
